+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [Version 1.5.1] - 2023-10-17
+
+- `MorphologicalFilterTask` adapted to work on boolean values.
+- Added `temporal_subset` method to `EOPatch`, which can be used to extract a subset of an `EOPatch` by filtering out temporal slices. Also added a corresponding `TemporalSubsetTask`.
+- `EOExecutor` now has an option to treat `TemporalDimensionWarning` as an exception.
+- String representation of `EOPatch` objects was revisited to avoid edge cases where the output would print enormous objects.
+
 ## [Version 1.5.0] - 2023-09-06
 
 The release focuses on making `eo-learn` much simpler to install, reducing the number of dependencies, and improving validation of soundness of `EOPatch` data.